requirements.txt
streamlit
torch==2.2.0
torchvision==0.17.0
opencv-python-headless
Pillow
numpy
scipy
seaborn
pyyaml
tqdm
matplotlib
requests
pandas
